backend/pyproject.toml
[project]
name = "placanje-core-backend"
version = "0.1.0"
description = "FastAPI backend for the Placanje product ecosystem"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "asyncpg>=0.30.0,<0.31.0",
  "certifi>=2025.1.31",
  "fastapi>=0.115.0,<0.116.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pyjwt[crypto]>=2.10.0,<3.0.0",
  "structlog>=25.0.0,<26.0.0",
  "uvicorn[standard]>=0.34.0,<0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<0.29.0",
  "mypy>=1.15.0,<2.0.0",
  "pytest-asyncio>=0.26.0,<0.27.0",
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.11.0,<0.12.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]

[tool.ruff.lint.per-file-ignores]
"app/api/v1/*.py" = ["B008"]
"app/core/auth.py" = ["B008"]
"app/core/deps.py" = ["B008"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
strict = true
ignore_missing_imports = true
disallow_untyped_defs = false
warn_return_any = false
